% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMBootstrapTest.R
\name{LMBootstrapTest}
\alias{LMBootstrapTest}
\title{Perform a test on the effects from the model}
\usage{
LMBootstrapTest(ResLMEffectMatrices, nboot = 100)
}
\arguments{
\item{ResLMEffectMatrices}{A list of 13 from \code{\link{LMEffectMatrices}}}

\item{nboot}{A integer with the number of iterations to perform (by default: 100)}
}
\value{
A list with the following elements:
\describe{
\item{\code{Fobs}}{A vector with the F statistics observed with the data}
\item{\code{Fboot}}{A matrix with the F statistics observed with the bootstrap}
\item{\code{Pvalues}}{A vector with the pvalue for every effect}
}
}
\description{
Compute a partial model without the tested effects then estimate the residuals. Next, compute new outcomes from the predicted values of the partial model and sampled residuals. Finally, compute the Sum of Squares and the test statistic.
}
\details{
The function works as follow:

To be written again
}
\examples{
 data('UCH')
 ResLMModelMatrix <- LMModelMatrix(formula=as.formula(UCH$formula),design=UCH$design)
 ResLMEffectMatrices <- LMEffectMatrices(ResLMModelMatrix = ResLMModelMatrix,outcomes=UCH$outcomes)

 result <- LMBootstrapTest(ResLMEffectMatrices = ResLMEffectMatrices,nboot=10)

}
\references{
Thiel M.,Feraud B. and Govaerts B. (2017) \emph{ASCA+ and APCA+: Extensions of ASCA and APCA
in the analysis of unbalanced multifactorial designs}, Journal of Chemometrics
}
