% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMModelMatrix.R
\name{LMModelMatrix}
\alias{LMModelMatrix}
\title{Creating the model matrix}
\usage{
LMModelMatrix(formula, design)
}
\arguments{
\item{formula}{The \emph{F} terms of the formula used to predict the response variables.\strong{Variable must be character}}

\item{design}{The \emph{nxk} "free encoded" experimental design data frame. n is the number of observations and k is number of factors(=main effects). \strong{Variables must be factors}.}
}
\value{
A list with the 6 following named elements :
\describe{
\item{\code{formula}}{A \code{\link{formula}} object with the expression of the model used to predict the responses.}
\item{\code{design}}{A \emph{nxk} data frame with the "free encoded" experimental design.}
\item{\code{ModelMatrix}}{A \emph{nxp} model matrix specifically encoded for the ASCA+ method. p is the number of model parameters with -1,0 or 1.}
\item{\code{ModelMatrixByEffect}}{A list with the model matrix for each of the \emph{F} terms of the formula}
\item{\code{covariateEffectsNames}}{A character vector with the name of the \emph{p} model parameters}
\item{\code{covariateEffectsNamesUnique}}{A character vector with the name of the \emph{F} model terms}
}
}
\description{
Create the specific model matrix from the design for the estimation of the model and the matrix decomposition.In a typical ASCA+ analysis, the model needs that the design is encoded with the \emph{sum coding}. Most of the times, the design matrix is a matrix with multiple factors that is not encoded with \emph{sum coding}. This function transforms the design into the correct model matrix.
}
\details{
Suppose the design matrix is \emph{nxk} with n observations and \emph{k} factors. After the re-encoding the model matrix
will be \emph{nxp}. For a parameter with \emph{a} levels the re-encoding is \emph{a-1} colums with 0 and 1 for \emph{a-1} first levels and -1 for the last.
\emph{p} is the number of columns for every factors from the design.
}
\examples{

data('UCH')

ResLMModelMatrix <- LMModelMatrix(as.formula(UCH$formula),UCH$design)

head(ResLMModelMatrix$ModelMatrix)

}
\references{
Thiel M.,Feraud B. and Govaerts B. (2017) \emph{ASCA+ and APCA+: Extensions of ASCA and APCA
in the analysis of unbalanced multifactorial designs}, Journal of Chemometrics
}
\seealso{
\code{\link{model.matrix}}

More informations about the specific encoding is available in the article from (\emph{Thiel et al}, 2017)
}
