% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMSS.R
\name{LMSS}
\alias{LMSS}
\title{Linear Model Sum of Squares}
\usage{
LMSS(ResLMEffectMatrices)
}
\arguments{
\item{ResLMEffectMatrices}{a ResLMEffectMatrices list from \code{\link{LMEffectMatrices}}}
}
\value{
A list with the following elements :
 \describe{
  \item{\code{formula}}{The \emph{p} terms formula of the GLM used to predict the response variables.}
  \item{\code{Type3Residuals}}{A list of matrices with the type III residuals for each model terms}
  \item{\code{variationPercentages}}{A \emph{p} named vector with the variation percentages of each model terms}
 }
}
\description{
Compute the type III sum of squares from a ResLMEffectMatrices list and return the variation percentages
}
\examples{
 data('UCH')
 ResLMModelMatrix <- LMModelMatrix(formula=as.formula(UCH$formula),design=UCH$design)
 ResLMEffectMatrices <- LMEffectMatrices(ResLMModelMatrix,outcomes=UCH$outcomes)
 ResLMSS <- LMSS(ResLMEffectMatrices)
 PlotLMSS(ResLMSS,abbrev=TRUE)

}
