% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCALMEffects.R
\name{PCALMEffects}
\alias{PCALMEffects}
\title{PCA on the effect matrices}
\usage{
PCALMEffects(ResLMEffectMatrices, method = c("ASCA", "APCA", "ASCA-E"))
}
\arguments{
\item{ResLMEffectMatrices}{a ResLMEffectMatrices list from \code{\link{LMEffectMatrices}}}

\item{method}{The method use to compute the PCA. One of \code{c("ASCA","APCA","ASCA-E")}}
}
\value{
A list of PCA results from \code{\link{SVDforPCA}} for each effect matrix. Those results contain :
 \describe{
  \item{\code{scores}}{Scores from the PCA for each component}
  \item{\code{loadings}}{Loadings from the PCA for each component}
  \item{\code{eigval}}{Eigenvalues of each component}
  \item{\code{pcd}}{Singular values}
  \item{\code{pcu}}{Normalized scores}
  \item{\code{var}}{Explained variance}
  \item{\code{cumvar}}{Cumulated explained variance}
  \item{\code{original.dataset}}{Original dataset}
 }
}
\description{
Run a Principal Component Analysis on the effect matrices
}
\examples{
 data('UCH')
 ResLMModelMatrix = LMModelMatrix(formula=as.formula(UCH$formula),design=UCH$design)
 ResLMEffectMatrices = LMEffectMatrices(ResLMModelMatrix,outcomes=UCH$outcomes)
 ResPCALMEffects = PCALMEffects(ResLMEffectMatrices,method="ASCA-E")
 DrawScores(ResPCALMEffects$Hippurate,type.obj="PCA")

}
