% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDesign.R
\name{PlotDesign}
\alias{PlotDesign}
\title{Plot the design matrix}
\usage{
PlotDesign(design, var1_name, var2_name, var3_name)
}
\arguments{
\item{design}{Design matrix of the data}

\item{var1_name}{Variable for the X axis}

\item{var2_name}{Variable for the Y axis}

\item{var3_name}{Variable used for the color, the size and the shape}
}
\value{
A ggplot of the design
}
\description{
Plot the experimental design of the data
}
\examples{
data(".RData")

}
\author{
Sebastien Franceschini
}
