% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLMSS.R
\name{PlotLMSS}
\alias{PlotLMSS}
\title{Barplot of the variation percentages}
\usage{
PlotLMSS(ResLMSS, abbrev = FALSE)
}
\arguments{
\item{ResLMSS}{A list from the \code{\link{LMSS}} function}

\item{abbrev}{If TRUE, the interaction names are abbreviated}
}
\value{
A barplot from ggplot2
}
\description{
Plotting the variation percentages of the model terms
}
\examples{
 data('UCH')
 ResLMModelMatrix = LMModelMatrix(formula=UCH$formula,design=UCH$design)
 ResLMEffectMatrices = LMEffectMatrices(ResLMModelMatrix,outcomes=UCH$outcomes)
 ResLMSS = LMSS(ResLMEffectMatrices)
 PlotLMSS(ResLMSS,abbrev=TRUE)

}
