% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLoadings.R
\name{PlotLoadings}
\alias{PlotLoadings}
\title{Plot the loadings from PCALMEffects}
\usage{
PlotLoadings(ResPCALMEffects, EffectNames, PCdim, ...)
}
\arguments{
\item{ResPCALMEffects}{A list from the function \code{\link{PCALMEffects}}}

\item{EffectNames}{A character vector of the p effects to plot}

\item{PCdim}{A vector of length p with the number of dimensions to draw for each effects}

\item{...}{More arguments to the function \code{\link{DrawLoadings}}}
}
\value{
A list of p graphs with the asked number of dimensions for each.
}
\description{
Plot the loadings from the ASCA, ASCA-E or APCA methods from PCALMEffects
}
