% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotScoresMatrix.R
\name{PlotScoresMatrix}
\alias{PlotScoresMatrix}
\title{Plotting a Scores Matrix}
\usage{
PlotScoresMatrix(ResPCALMEffects, design, EffectNames = NULL, alleffect,
  PCdim = NULL, varname.colorup = NULL, varname.colordown = NULL,
  varname.pch = NULL, vec.colorup, vec.colordown, vec.pch)
}
\arguments{
\item{ResPCALMEffects}{object}

\item{design}{design}

\item{EffectNames}{name}

\item{alleffect}{logical}

\item{PCdim}{pc}

\item{varname.colorup}{varname}

\item{varname.colordown}{varname}

\item{varname.pch}{varname}

\item{vec.colorup}{vec}

\item{vec.colordown}{vec}

\item{vec.pch}{vec}
}
\value{
A graph
}
\description{
Plot a matrix of scores graphs
}
