% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotScoresXY.R
\name{PlotScoresXY}
\alias{PlotScoresXY}
\title{Plotting the scores from the PCALMEffects function}
\usage{
PlotScoresXY(ResPCALMEffects, design, EffectVector, PCaxes = c(1, 2),
  varname.color, varname.pch, ...)
}
\arguments{
\item{ResPCALMEffects}{A list of p elements depending of the model terms from \code{\link{PCALMEffects}}}

\item{design}{A n x d "free encoded" experimental design matrix with n observations and d the categorical variables.}

\item{EffectVector}{A character vector of length l with the name of the model terms to plot}

\item{PCaxes}{A vector with the two PC to plot for every effects}

\item{varname.color}{A character variable with the name of the model term to use as color}

\item{varname.pch}{A character variable with the name of the model term to use as pch}

\item{...}{Other arguments from \code{\link{DrawScores}}}
}
\value{
A list of l ggplot2 graphs
}
\description{
Draw one or more graphs of the scores from a PCA. More specifically the result of the \code{\link{PCALMEffects}} function.
}
\details{
Some messages from ggplot2 could be suppressed while we overwrite the graphs.
}
\examples{
 data('UCH')
 ResLMModelMatrix = LMModelMatrix(formula=as.formula(UCH$formula),design=UCH$design)
 ResLMEffectMatrices = LMEffectMatrices(ResLMModelMatrix,outcomes=UCH$outcomes)
 ResPCALMEffects = PCALMEffects(ResLMEffectMatrices,method="ASCA-E")
 PlotScoresXY(ResPCALMEffects = ResPCALMEffects, design = UCH$design,
  EffectVector = c("Time","Hippurate"),varname.color = "Citrate", varname.pch = "Hippurate")

}
