% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotScree.R
\name{PlotScree}
\alias{PlotScree}
\title{Plot a Screeplot from a ResPCALMEffects list}
\usage{
PlotScree(ResPCALMEffects, EffectName, PCnumber = 5)
}
\arguments{
\item{ResPCALMEffects}{a ResPCALMEffects list from the function \code{\link{PCALMEffects}}}

\item{EffectName}{a character with the name of the effect to plot}

\item{PCnumber}{Integer with the number of components to plot}
}
\value{
A ggplot2 object with the Screeplot
}
\description{
Plot a Screeplot of the Principal Component variances
}
