% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEffectMatrices.R
\name{LMEffectMatrices}
\alias{LMEffectMatrices}
\title{Computing the Effect Matrices}
\usage{
LMEffectMatrices(ResLMModelMatrix, outcomes)
}
\arguments{
\item{ResLMModelMatrix}{A list of 3 from \code{\link{LMModelMatrix}}}

\item{outcomes}{A (n x y) matrix with n observations and y response variables}
}
\value{
A list with the following elements:
 \describe{
   \item{\code{formula}}{}
   \item{\code{design}}{}
   \item{\code{ModelMatrix}}{}
   \item{\code{outcomes}}{}
   \item{\code{effectMatrices}}{}
   \item{\code{modelMatrixByEffect}}{}
   \item{\code{predictedvalues}}{}
   \item{\code{residuals}}{}
   \item{\code{parameters}}{}
   \item{\code{covariateEffectsNamesUnique}}{}
   \item{\code{covariateEffectsNames}}{}
 }

 @examples
 data('UCH')
 ResLMModelMatrix = LMModelMatrix(formula=UCH$formula,design=UCH$design)
 LMEffectMatrices(ResLMModelMatrix,outcomes=UCH$outcomes)

 @import stringr
 @import plyr
}
\description{
Runs a GLM model and decomposes the outcomes into effect matrices
}
