% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMModelMatrix.R
\name{LMModelMatrix}
\alias{LMModelMatrix}
\title{Creating the model matrix}
\usage{
LMModelMatrix(formula, outcomes, design)
}
\description{
Create the specific model matrix needed for the ASCA+ method from the design and model matrices
}
