% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMModelMatrix.R
\name{LMModelMatrix}
\alias{LMModelMatrix}
\title{Creating the model matrix}
\usage{
LMModelMatrix(formula, design)
}
\arguments{
\item{formula}{The \emph{p} terms formula of the GLM used to predict the response variables.\strong{Variable must be character}}

\item{design}{The nxk "free encoded" experimental design data frame. \strong{Variables must be factors}.}
}
\value{
A list with the 3 following named elements :
\describe{
\item{\code{formula}}{A \code{\link{formula}} object with the expression of the GLM used to predict the outcomes.}
\item{\code{design}}{A \emph{nxk} data frame with the "free encoded" experimental design.}
\item{\code{ModelMatrix}}{A \emph{nxK} model matrix specifically encoded for the ASCA-GLM method.}
}
}
\description{
Create the specific model matrix from the design for the ASCA-GLM method.
}
\details{
In typical ASCA-glm analysis, the GLM model need the design to be re-encoded with the commonly used in industrial experimental design \emph{sum coding}. The result is the model matrix.
Suppose the design matrix is \emph{nxk} with n observations and \emph{k} factors. After the transformation the model matrix
will be \emph{nxK}. For a parameter with a levels the re-encoding is \emph{a-1} colums with 0 and 1 for \emph{a-1} first levels and -1 for the last.
\emph{K} is the sum of all the columns for every parameters.
}
\examples{

data('UCH')

ResLMModelMatrix <- LMModelMatrix(as.formula(UCH$formula),UCH$design)

head(ResLMModelMatrix$ModelMatrix)

}
\references{
Thiel M.,Feraud B. and Govaerts B. (2017) \emph{ASCA+ and APCA+: Extensions of ASCA and APCA
in the analysis of unbalanced multifactorial designs}, Journal of Chemometrics
}
\seealso{
\code{\link{model.matrix}}

More informations about the specific encoding is available in the article from (\emph{Thiel et al}, 2017)
}
