% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMSS.R
\name{LMSS}
\alias{LMSS}
\title{Linear Model Sum of Squares}
\usage{
LMSS(ResLMEffectMatrices)
}
\arguments{
\item{ResLMEffectMatrices}{a ResLMEffectMatrices list from \code{\link{LMEffectMatrices}}}
}
\value{
A list with the following elements :
 \describe{
  \item{\code{formula}}{A formula object with the expression of the GLM used to predict the outcomes}
  \item{\code{Type3Residuals}}{}
  \item{\code{variationPercentages}}{}
 }
}
\description{
Compute the type III sum of squares from a ResLMEffectMatrices list and return the variation percentages
}
\examples{
 data('UCH')
 ResLMModelMatrix = LMModelMatrix(formula=UCH$formula,design=UCH$design)
 ResLMEffectMatrices = LMEffectMatrices(ResLMModelMatrix,outcomes=UCH$outcomes)
 LMSS(ResLMEffectMatrices)

}
