% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLMSS.R
\name{PlotLMSS}
\alias{PlotLMSS}
\title{Barplot of the variation percentages}
\usage{
PlotLMSS(ResLMEffectMatrices, abbrev = FALSE)
}
\arguments{
\item{ResLMEffectMatrices}{A list from the \code{\link{LMEffectMatrices}} function}

\item{abbrev}{Logical, if TRUE, the interaction names are abbreviated}
}
\value{
A barplot from ggplot2
}
\description{
Plotting the variation percentages of the model terms
}
\examples{
data('UCH')
ResLMModelMatrix = LMModelMatrix(formula=as.formula(UCH$formula),design=UCH$design)
ResLMEffectMatrices = LMEffectMatrices(ResLMModelMatrix,outcomes=UCH$outcomes)
PlotLMSS(ResLMEffectMatrices,abbrev=TRUE)

}
