% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotScoresMatrix.R
\name{PlotScoresMatrix}
\alias{PlotScoresMatrix}
\title{Plotting a Scores Matrix}
\usage{
PlotScoresMatrix(ResPCALMEffects, ModelAbbrev = FALSE, design,
  EffectNames = NULL, alleffect, PCdim = NULL,
  varname.colorup = NULL, varname.colordown = NULL,
  varname.pchup = NULL, varname.pchdown = NULL, vec.colorup = NULL,
  vec.colordown = NULL, vec.pchup = NULL, vec.pchdown = NULL)
}
\arguments{
\item{ResPCALMEffects}{A list of p+3 elements depending of the model terms from \code{\link{PCALMEffects}}}

\item{ModelAbbrev}{A logical whether to abbreviate the interaction terms}

\item{design}{The nxk "free encoded" experimental design data frame}

\item{EffectNames}{A character vector with the name of the effects to plot}

\item{alleffect}{A logical whether to plot every effect}

\item{PCdim}{A numeric vector with the same length than EffectNames and indicating the number of component to plot}

\item{varname.colorup}{A character with the name of variable used to color the upper triangle}

\item{varname.colordown}{A character with the name of variable used to color the upper triangle}

\item{varname.pchup}{A character with the name of variable used to mark points from the upper triangle}

\item{varname.pchdown}{A character with the name of variable used to mark points from the lower triangle}

\item{vec.colorup}{A color vector with a length equivalent to the number of levels from varname.colorup}

\item{vec.colordown}{A color vector with a length equivalent to the number of levels from varname.colordown}

\item{vec.pchup}{A pch vector with a length equivalent to the number of levels from varname.pchup}

\item{vec.pchdown}{A pch vector with a length equivalent to the number of levels from varname.pchdown}
}
\value{
A matrix of graphs
}
\description{
Plot a matrix of scores graphs
}
\examples{

 data('UCH')
 ResLMModelMatrix = LMModelMatrix(formula=as.formula(UCH$formula),design=UCH$design)
 ResLMEffectMatrices = LMEffectMatrices(ResLMModelMatrix,outcomes=UCH$outcomes)
 ResPCALMEffects = PCALMEffects(ResLMEffectMatrices,method="ASCA-E")
 PlotScoresMatrix(ResPCALMEffects,
                 ModelAbbrev=FALSE,
                 design=UCH$design,
                 EffectNames = c("Citrate","Hippurate","Hippurate:Citrate"),
                 alleffect = FALSE,
                 PCdim=c(2,2,2),
                 varname.colorup = "Citrate",
                 vec.colorup = c("red","blue","green"),
                 varname.pchup="Hippurate",
                 vec.pchup=c(1,2,3),
                 varname.pchdown = "Day",
                 vec.pchdown = c(4,5),
                 varname.colordown="Time",
                 vec.colordown = c("brown","grey"))

}
